requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
pymongo==4.6.0
motor==3.3.2

# AI/ML
openai==1.20.0
python-dotenv==1.0.0

# Utilities
httpx==0.25.2
tenacity==8.2.3
python-dateutil==2.8.2
tzdata==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# RAG / Vector Database
chromadb==0.4.22
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
sentence-transformers==2.2.2

# Dashboard
streamlit==1.29.0

# Telegram Bot
python-telegram-bot==20.7

# Security
PyJWT==2.8.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6

# Additional utilities for production
slowapi==0.1.9
sentry-sdk[fastapi]==1.39.1
